SystemMonitor: Remember visible columns
Save the visible columns in the config file of SystemMonitor.
The visible columns are stored as a comma separated list of the names
corresponding to the ProcessModel::Column enum. If no value is given
per default the columns PID, TID, Name, CPU, User, Virtual and
DirtyPrivate are set to be visible.